+++ src/features/cardPreview/constants/CardPreview.constant.ts
@@ -1,0 +1,7 @@
+export const CARD_IDENTIFYING_NUMBER = {
+  VISA: 4,
+  MASTERCARD: {
+    MIN: 51,
+    MAX: 55,
+  },
+};
