+++ src/features/cardPreview/styles/CardPreview.css.ts
