+++ src/features/cardPreview/types/CardPreview.types.ts
@@ -1,0 +1,11 @@
+import {
+  CardNumberPosition,
+  ExpirationPeriod,
+  CardTypeList,
+} from "../../../types/index.types";
+
+export type CardPreviewProps = {
+  cardNumber: Record<CardNumberPosition, string>;
+  expirationPeriod: Record<ExpirationPeriod, string>;
+  cardType: keyof CardTypeList | "";
+};
