+++ src/features/cardPreview/ui/CardPreview.tsx
@@ -6,54 +6,17 @@ import {
   StyledCardNumberWrap,
   StyledCardNumber,
   StyledExpirationPeriod,
-} from "./CardPreview.css";
-import { CardTypeList } from "../../types/index.types";
-import { CardNumberPosition, ExpirationPeriod } from "../../types/index.types";
-import { INITIALIZE_VALUE } from "../../shared/constants/constant";
-import { useMemo } from "react";
-
-type CardNumber = Record<CardNumberPosition, string>;
-
-type CardPreviewProps = {
-  cardNumber: CardNumber;
-  expirationPeriod: Record<ExpirationPeriod, string>;
-  cardType: keyof CardTypeList | "";
-};
-
-const CARD_IDENTIFYING_NUMBER = {
-  VISA: 4,
-  MASTERCARD: {
-    MIN: 51,
-    MAX: 55,
-  },
-};
-
-function getIdentifyFns(id: string) {
-  return [
-    {
-      identify: () => Number(id[0]) === CARD_IDENTIFYING_NUMBER.VISA,
-      logoSrc: "./images/Visa.svg",
-    },
-    {
-      identify: () =>
-        Number(id) >= CARD_IDENTIFYING_NUMBER.MASTERCARD.MIN &&
-        Number(id) <= CARD_IDENTIFYING_NUMBER.MASTERCARD.MAX,
-      logoSrc: "./images/Mastercard.svg",
-    },
-  ];
-}
+} from "../styles/CardPreview.css";
+import { CardPreviewProps } from "../types/CardPreview.types";
+import { INITIALIZE_VALUE } from "../../../shared/constants/constant";
+import useLogoSrc from "../hooks/useLogoSrc";
 
 function CardPreview({
   cardNumber,
   expirationPeriod,
   cardType,
 }: CardPreviewProps) {
-  const logoSrc = useMemo(() => {
-    const id = cardNumber.first.slice(0, 2);
-    const identifyFns = getIdentifyFns(id);
-    const identifiedLogoSrc = identifyFns.find((fn) => fn.identify());
-    return identifiedLogoSrc?.logoSrc ?? INITIALIZE_VALUE;
-  }, [cardNumber]);
+  const { logoSrc } = useLogoSrc(cardNumber);
 
   return (
     <StyledContainer cardType={cardType}>
